pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ruvox"
version = "0.1.0"
description = "RuVox — Russian text preprocessing pipeline for TTS Silero"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "evgen" }
]
dependencies = [
    "evdev-binary>=1.9.3,<2",
    "num2words>=0.5.12,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=4.0,<6",
]
g2p = [
    "g2p-en>=2.1.0,<3",
    "eng-to-ipa>=0.0.2,<1",
]
tts = [
    "torch>=2.0,<3",
    "scipy>=1.10,<2",
    "numpy>=1.24,<3",
    "omegaconf>=2.3,<3",
]
ui = [
    "PyQt6>=6.6.0,<7",
    "PyQt6-WebEngine>=6.6.0,<7",  # QWebEngineView for Mermaid diagram rendering
    "markdown>=3.5,<4",
    "dasbus>=1.7,<2",  # D-Bus library for global hotkeys (requires PyGObject)
    "python-mpv>=1.0,<2",  # libmpv bindings for audio playback (scaletempo2)
]

[project.scripts]
ruvox = "ruvox.ui.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ruvox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/ruvox"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false

[dependency-groups]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=4.0,<6",
]
